% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{attribute_bijection}
\alias{attribute_bijection}
\title{Constructs Unique Attribute Pattern Map}
\usage{
attribute_bijection(K)
}
\arguments{
\item{K}{Number of Attributes.}
}
\value{
A \code{vec} with length \eqn{K} detailing the power's of 2.
}
\description{
Computes the powers of 2 from \eqn{0} up to \eqn{K - 1} for
\eqn{K}-dimensional attribute pattern.
}
\examples{
## Construct an attribute bijection ----
biject = attribute_bijection(3)
}
\seealso{
\code{\link[simcdm:attribute_inv_bijection]{simcdm::attribute_inv_bijection()}}
}
\author{
Steven Andrew Culpepper and James Joseph Balamuta
}
