% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bijectionvector}
\alias{bijectionvector}
\title{Bijection Vector}
\usage{
bijectionvector(K)
}
\arguments{
\item{K}{Number of Attributes.}
}
\value{
A \code{vec} with length \eqn{K} detailing the power's of 2.
}
\description{
Computes the powers of 2 from \eqn{0} up to \eqn{K - 1}.
}
\examples{

bijectionvector(3)

}
