% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_bijectionvector}
\alias{inv_bijectionvector}
\title{Perform an Inverse Bijection of an Integer to Attribute Pattern}
\usage{
inv_bijectionvector(K, CL)
}
\arguments{
\item{K}{Number of Attributes.}

\item{CL}{An \code{integer} between \eqn{0} and \eqn{2^{K-1}}}
}
\value{
A \eqn{K}-dimensional vector with an attribute pattern corresponding
to \code{CL}.
}
\description{
Convert integer between \eqn{0} and \eqn{2^{K-1}} to
\eqn{K}-dimensional attribute pattern.
}
\examples{
inv_bijectionvector(5, 1)
inv_bijectionvector(5, 2)
}
