% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_reference.R
\name{pi_reference}
\alias{pi_reference}
\title{Mapping of Entries of Pi to Latent Attribute Classes}
\usage{
pi_reference(K)
}
\arguments{
\item{K}{A \code{numeric}, denoting the number of attributes.}
}
\value{
A \code{matrix}, where row c represents the latent class
corresponding to entry c of pi
}
\description{
Obtains mapping of entries of pi to latent attribute classes.
}
\author{
Steven Andrew Culpepper, Aaron Hudson
}
