% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_alpha_matrix}
\alias{sim_alpha_matrix}
\title{Simulate the Latent Attribute Profile Matrix \eqn{\mathbf{\alpha}_c}}
\usage{
sim_alpha_matrix(K)
}
\arguments{
\item{K}{Number of Skills}
}
\value{
A \eqn{2^K} by \eqn{K} \code{matrix} of latent classes
corresponding to entry \eqn{c} of \eqn{pi} based upon
mastery and nonmastery of the \eqn{K} skills.
}
\description{
Generate the \eqn{\mathbf{\alpha}_c = (\alpha_{c1}, \ldots, \alpha_{cK})'}
attribute profile matrix for members of class \eqn{c} such that \eqn{\alpha_{ck}}
is 1 if members of class \eqn{c} possess skill \eqn{k} and zero otherwise.
}
\examples{
# Define test parameters and traits
K = 3

# Generate an Latent Attribute Profile (Alpha) Matrix
alphas = sim_alpha_matrix(K)
}
\author{
James Joseph Balamuta and Steven Andrew Culpepper
}
