% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_dina}
\alias{sim_dina}
\title{Simulation Responses from the DINA model}
\usage{
sim_dina(alphas, Q, ss, gs)
}
\arguments{
\item{alphas}{A \eqn{N} by K \code{matrix} of latent attributes.}

\item{Q}{A \eqn{N} by K \code{matrix} indicating which skills are required for which items.}

\item{ss}{A \eqn{J} \code{vector} of item slipping parameters.}

\item{gs}{A \eqn{J} \code{vector} of item guessing parameters.}
}
\value{
A \eqn{N} by \eqn{J} \code{matrix} of responses from the DINA model.
}
\description{
Sample responses from the DINA model for given attribute profiles, Q matrix,
and item parmeters. Returns a \code{matrix} of dichotomous responses
generated under DINA model.
}
\examples{
###########################################
#de la Torre (2009) Simulation Replication
###########################################
N = 200
K = 5
J = 30
delta0 = rep(1,2^K)
    
# Creating Q matrix
Q = matrix(rep(diag(K),2),2*K,K,byrow=TRUE)
for(mm in 2:K){
    temp = combn(1:K,m=mm)
    tempmat = matrix(0,ncol(temp),K)
    for(j in 1:ncol(temp)) tempmat[j,temp[,j]] = 1
    Q = rbind(Q,tempmat)
}
Q = Q[1:J,]
        
# Setting item parameters and generating attribute profiles
ss = gs = rep(.2,J)
PIs = rep(1/(2^K),2^K)
CLs = c((1:(2^K))\\\%*\\\%rmultinom(n=N,size=1,prob=PIs) )
            
# Defining matrix of possible attribute profiles
As = rep(0,K)
for(j in 1:K){
    temp = combn(1:K,m=j)
    tempmat = matrix(0,ncol(temp),K)
    for(j in 1:ncol(temp)) tempmat[j,temp[,j]] = 1
    As = rbind(As,tempmat)
}
As = as.matrix(As)
                    
# Sample true attribute profiles
Alphas = As[CLs,]
                    
# Simulate data under DINA model 
gen = sim_dina_items(Alphas,Q,ss,gs)
Y_sim = gen$Y
}
\seealso{
\code{\link{DINA_Gibbs} }
}
\author{
Steven Andrew Culpepper
}
