% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_dina_items}
\alias{sim_dina_items}
\title{Simulation Responses from the DINA model}
\usage{
sim_dina_items(alphas, Q, ss, gs)
}
\arguments{
\item{alphas}{A \eqn{N} by \eqn{K} \code{matrix} of latent attributes.}

\item{Q}{A \eqn{J} by \eqn{K} \code{matrix} indicating which skills are
required for which items.}

\item{ss}{A \eqn{J} \code{vector} of item slipping parameters.}

\item{gs}{A \eqn{J} \code{vector} of item guessing parameters.}
}
\value{
A \eqn{N} by \eqn{J} \code{matrix} of responses from the DINA model.
}
\description{
Sample responses from the DINA model for given attribute profiles, Q matrix,
and item parmeters. Returns a \code{matrix} of dichotomous responses
generated under DINA model.
}
\examples{
N = 200
K = 5
J = 30
delta0 = rep(1, 2 ^ K)

# Creating Q matrix
Q = matrix(rep(diag(K), 2), 2 * K, K, byrow = TRUE)
for (mm in 2:K) {
  temp = combn(seq_len(K), m = mm)
  tempmat = matrix(0, ncol(temp), K)
  for (j in seq_len(ncol(temp)))
    tempmat[j, temp[, j]] = 1
  Q = rbind(Q, tempmat)
}
Q = Q[seq_len(J), ]

# Setting item parameters and generating attribute profiles
ss = gs = rep(.2, J)
PIs = rep(1 / (2 ^ K), 2 ^ K)
CLs = c((1:(2 ^ K)) \%*\% rmultinom(n = N, size = 1, prob = PIs))

# Defining matrix of possible attribute profiles
As = rep(0, K)
for (j in seq_len(K)) {
  temp = combn(1:K, m = j)
  tempmat = matrix(0, ncol(temp), K)
  for (j in seq_len(ncol(temp)))
    tempmat[j, temp[, j]] = 1
  As = rbind(As, tempmat)
}
As = as.matrix(As)

# Sample true attribute profiles
Alphas = As[CLs, ]

# Simulate item data under DINA model 
dina_items = sim_dina_items(Alphas, Q, ss, gs)

# Simulate attribute data under DINA model 
dina_attributes = sim_dina_attributes(Alphas, Q)
}
\seealso{
\code{\link[simcdm:sim_dina_class]{simcdm::sim_dina_class()}} and \code{\link[simcdm:sim_dina_attributes]{simcdm::sim_dina_attributes()}}
}
\author{
Steven Andrew Culpepper and James Joseph Balamuta
}
