% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_eta_matrix}
\alias{sim_eta_matrix}
\title{Generate ideal response \eqn{\eta} Matrix}
\usage{
sim_eta_matrix(K, J, Q)
}
\arguments{
\item{K}{Number of Attribute Levels}

\item{J}{Number of Assessment Items}

\item{Q}{Q Matrix with dimensions \eqn{K \times J}{K x J}.}
}
\value{
A \code{mat} with dimensions \eqn{J \times 2^K}{J x 2^K}.
}
\description{
Creates the ideal response matrix for each trait
}
\examples{
# Fixed Number of Assessment Items for Q
J = 18

# Fixed Number of Attributes for Q
K       = 3

# Specify Q
qbj = c(4, 2, 1, 4, 2, 1, 4, 2, 1, 6, 5, 3, 6, 5, 3, 7, 7, 7)

# Fill Q Matrix
Q = matrix(, J, K)
for (j in seq_len(J)) {
  Q[j,] = inv_bijectionvector(K, qbj[j])
}

# Create an eta matrix
ETA = sim_eta_matrix(K, J, Q)

# Generate an ETA matrix for a random Q.

# Create an eta matrix
Q_sim = sim_q_matrix(K, J)
ETA_gen = sim_eta_matrix(K, J, Q_sim)
}
