% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_q_matrix}
\alias{sim_q_matrix}
\title{Generate random Q matrix}
\usage{
sim_q_matrix(J, K)
}
\arguments{
\item{J}{Number of Items}

\item{K}{Number of Attributes}
}
\value{
A dichotomous \code{matrix} for Q.
}
\description{
Simulates a Q matrix containing three identity matrices after a row
permutation.
}
\examples{
sim_q_matrix(7, 2)

sim_q_matrix(10, 3)
}
