% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcdm-package.R
\docType{package}
\name{simcdm-package}
\alias{simcdm}
\alias{simcdm-package}
\title{simcdm: Simulate Cognitive Diagnostic Model (CDM) Data}
\description{
Provides efficient R and 'C++' routines to simulate cognitive diagnostic
model data for Deterministic Input, Noisy "And" Gate (DINA) and
reduced Reparametrized Unifed Model (rRUM).
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

Authors:
\itemize{
  \item Steven Culpepper \email{sculpepp@illinois.edu}
}

}
