% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcdm-package.R
\docType{package}
\name{simcdm-package}
\alias{simcdm}
\alias{simcdm-package}
\title{simcdm: Simulate Cognitive Diagnostic Model ('CDM') Data}
\description{
Provides efficient R and 'C++' routines to simulate cognitive diagnostic model data for Deterministic Input, Noisy "And" Gate ('DINA') and reduced Reparameterized Unified Model ('rRUM') from Culpepper and Hudson (2017) \doi{ 10.1177/0146621617707511}, Culpepper (2015) \doi{10.3102/1076998615595403}, and de la Torre (2009) \doi{10.3102/1076998607309474}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tmsalab.github.io/simcdm/}
  \item \url{https://github.com/tmsalab/simcdm}
  \item Report bugs at \url{https://github.com/tmsalab/simcdm/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Aaron Hudson \email{awhudson@uw.edu} (\href{https://orcid.org/0000-0002-9731-2224}{ORCID}) [contributor, copyright holder]
}

}
